- refactor of FloatCanvas package, unittest and demo run on Py27 and Py33
