--- wx/lib/floatcanvas/SpecialObjects/__init__.py
@@ -1,7 +1,0 @@
-"""
-SpecialObjects Package
-
-Various special objects -- to specific for inclusion in the main FloatCanvas
-
-"""
-from PieChart import PieChart
